% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course functions.R
\name{makesliders}
\alias{makesliders}
\title{Create a sliders file}
\usage{
makesliders(curves, id, begin = NULL, end = NULL)
}
\arguments{
\item{curves}{Vector indicating curves}

\item{id}{Curve names}

\item{begin}{First static landmark number}

\item{end}{Last static landmark number}
}
\value{
Sliders object
}
\description{
Make a sliders file from a character vector indicating which landmarks are curves
}
