% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course functions.R
\name{model.space}
\alias{model.space}
\title{Model 2 dimensional shape from PC scores}
\usage{
model.space(PCscores, dim, eigenvectors, consensus)
}
\arguments{
\item{PCscores}{scores}

\item{dim}{dimensions}

\item{eigenvectors}{eigenvectors}

\item{consensus}{consensus}
}
\value{
shape plot
}
\description{
Model 2 dimensional shape from PC scores
}
