% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{shrewteeth}
\alias{shrewteeth}
\title{Shrewteeth}
\format{TPS file}
\description{
A dataset containing dental 2D landmarks from photographs of European shrews
}
\details{
Associated photos can be found in 'extdata/ShrewsAndMarmots'
}
\references{
Polly, P. D. 2003. Paleophylogeography of Sorex araneus: molar shape as a morphological marker for fossil shrews. Mammalia, 68: 233-243
}
