% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/course functions.R
\name{subsetgeom}
\alias{subsetgeom}
\title{Subset geomorph data frames}
\usage{
subsetgeom(gm.data.frame, dim, value, keep)
}
\arguments{
\item{gm.data.frame}{geomorph data frame}

\item{dim}{land or spec}

\item{value}{vector with which items to remove}

\item{keep}{T or F - keep or remove values}
}
\value{
subsampled geomorph data frame
}
\description{
Allows subsetting of geomorph data frames (and all variables therein) either by specimen or by landmark
}
\details{
The specified values may be either kept or removed from the final object
}
