% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tarsalsdf}
\alias{tarsalsdf}
\title{Tarsalsdf}
\format{csv file}
\description{
Geomorph dataframe containing 2D and 3D landmarks for carnivoran tarsals and
associated ecological data
}
\details{
From Polly XXX [cite]
}
